-Reverted some changes to make the code build again on Tiger. -Removed x86_64 from Deployment configuration. -macosx: Use SDL for sound plugin, removed Carbon backend. -(MaddTheSane)Fixed memory leaks (Patch #8427).
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@63548 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ macosx/PcsxPlugin.m
@@ -274,7 +274,7 @@
 	return path;
 }
 
-- (NSUInteger)hash
+- (unsigned int)hash
 {
     return [path hash];
 }
